requirements.txt
# dev tools
black

streamlit
pandas
openpyxl

# Google Drive API
google-api-python-client
google-auth-httplib2
google-auth-oauthlib